@@ -58,41 +56,29 @@ IF(MSVC_ARM64)
  SKIP_ROCKSDB_PLUGIN("Windows ARM64 not supported")
 ENDIF()
 
-#
-# Also, disable on ARM64 when not Linux
-# Requires submodule update to v6.16.3
-# containing commit https://github.com/facebook/rocksdb/commit/ee4bd4780b321ddb5f92a0f4eb956f2a2ebd60dc
-#
-IF(CMAKE_SYSTEM_PROCESSOR MATCHES "(arm64|aarch64)" AND NOT CMAKE_SYSTEM_NAME STREQUAL "Linux")
-  SKIP_ROCKSDB_PLUGIN("ARM64 disabled on all except Linux")
-ENDIF()
-
-# This plugin needs recent C++ compilers (it is using C++11 features)
+# This plugin needs recent C++ compilers (it is using C++17 features)
 # Skip build for the old compilers
-SET(CXX11_FLAGS)
-SET(OLD_COMPILER_MSG "requires c++11 -capable compiler (minimal supported versions are g++ 4.8, clang 3.3, VS2015)")
+SET(CXX17_FLAGS)
+SET(OLD_COMPILER_MSG "requires c++17 -capable compiler (minimal supported versions are g++ 7, clang 5, VS2017)")
 
 IF(CMAKE_CXX_COMPILER_ID MATCHES "GNU")
   EXECUTE_PROCESS(COMMAND ${CMAKE_CXX_COMPILER} -dumpversion OUTPUT_VARIABLE GCC_VERSION)
-  IF (GCC_VERSION VERSION_LESS 4.8)
+  IF (GCC_VERSION VERSION_LESS 7)
     SKIP_ROCKSDB_PLUGIN("${OLD_COMPILER_MSG}")
   ENDIF()
-  SET(CXX11_FLAGS "-std=c++11")
-  IF (GCC_VERSION VERSION_LESS 5.0)
-    SET(CXX11_FLAGS "-std=c++11 -Wno-missing-field-initializers")
-  ENDIF()
+  SET(CXX17_FLAGS "-std=c++17")
 ELSEIF (CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  IF ((CMAKE_CXX_COMPILER_VERSION AND CMAKE_CXX_COMPILER_VERSION VERSION_LESS  3.3) OR
-     (CLANG_VERSION_STRING AND CLANG_VERSION_STRING VERSION_LESS  3.3))
+  IF ((CMAKE_CXX_COMPILER_VERSION AND CMAKE_CXX_COMPILER_VERSION VERSION_LESS  5) OR
+     (CLANG_VERSION_STRING AND CLANG_VERSION_STRING VERSION_LESS  5))
     SKIP_ROCKSDB_PLUGIN("${OLD_COMPILER_MSG}")
   ENDIF()
-  SET(CXX11_FLAGS "-std=c++11 -stdlib=libstdc++")
+  SET(CXX17_FLAGS "-std=c++17 -stdlib=libstdc++")
   IF(MSVC)
     # clang-cl does not work yet
     SKIP_ROCKSDB_PLUGIN("Clang-cl is not supported")
   ENDIF()
 ELSEIF(MSVC)
-  IF (MSVC_VERSION LESS 1900)
+  IF (MSVC_VERSION LESS 1914)
     SKIP_ROCKSDB_PLUGIN("${OLD_COMPILER_MSG}")
   ENDIF()
 ELSE()
@@ -100,9 +86,9 @@ ELSE()
 ENDIF()
 
 IF(CMAKE_VERSION GREATER 3.0)
-  SET(CMAKE_CXX_STANDARD  11)
-ELSEIF(CXX11_FLAGS)
-  ADD_DEFINITIONS(${CXX11_FLAGS})
+  SET(CMAKE_CXX_STANDARD  17)
+ELSEIF(CXX17_FLAGS)
+  ADD_DEFINITIONS(${CXX17_FLAGS})
 ENDIF()
 
 SET(ROCKSDB_SE_SOURCES
