@@ -24,6 +24,10 @@
 #include <my_global.h>
 
 #include <inttypes.h>
+#include <string>
+#include <vector>
+
+#include "./debug_sync.h"
 
 /* This C++ files header file */
 #include "./rdb_cf_manager.h"
@@ -33,8 +37,7 @@
 #include "./ha_rocksdb_proto.h"
 #include "./rdb_datadic.h"
 #include "./rdb_psi.h"
-
-#include <string>
+#include "./rdb_threads.h"
 
 namespace myrocks {
 
@@ -56,17 +59,18 @@ void Rdb_cf_manager::init(
 
   m_cf_options = std::move(cf_options);
 
-  for (auto cfh : *handles) {
-    DBUG_ASSERT(cfh != nullptr);
-    m_cf_name_map[cfh->GetName()] = cfh;
-    m_cf_id_map[cfh->GetID()] = cfh;
+  for (auto cfh_ptr : *handles) {
+    DBUG_ASSERT(cfh_ptr != nullptr);
+
+    std::shared_ptr<rocksdb::ColumnFamilyHandle> cfh(cfh_ptr);
+    m_cf_name_map[cfh_ptr->GetName()] = cfh;
+    m_cf_id_map[cfh_ptr->GetID()] = cfh;
   }
 }
 
 void Rdb_cf_manager::cleanup() {
-  for (auto it : m_cf_name_map) {
-    delete it.second;
-  }
+  m_cf_name_map.clear();
+  m_cf_id_map.clear();
   mysql_mutex_destroy(&m_mutex);
   m_cf_options = nullptr;
 }
@@ -78,20 +82,17 @@ void Rdb_cf_manager::cleanup() {
   @detail
     See Rdb_cf_manager::get_cf
 */
-rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_or_create_cf(
-    rocksdb::DB *const rdb, const std::string &cf_name_arg) {
+std::shared_ptr<rocksdb::ColumnFamilyHandle> Rdb_cf_manager::get_or_create_cf(
+    rocksdb::DB *const rdb, const std::string &cf_name) {
   DBUG_ASSERT(rdb != nullptr);
+  DBUG_ASSERT(!cf_name.empty());
+  std::shared_ptr<rocksdb::ColumnFamilyHandle> cf_handle;
 
-  rocksdb::ColumnFamilyHandle *cf_handle = nullptr;
-
-  if (cf_name_arg == PER_INDEX_CF_NAME) {
+  if (cf_name == PER_INDEX_CF_NAME) {
     // per-index column families is no longer supported.
     my_error(ER_PER_INDEX_CF_DEPRECATED, MYF(0));
-    return nullptr;
+    return cf_handle;
   }
-
-  const std::string &cf_name =
-      cf_name_arg.empty() ? DEFAULT_CF_NAME : cf_name_arg;
 
   RDB_MUTEX_LOCK_CHECK(m_mutex);
 
@@ -114,14 +115,15 @@ rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_or_create_cf(
     sql_print_information("    target_file_size_base=%" PRIu64,
                           opts.target_file_size_base);
 
+    rocksdb::ColumnFamilyHandle *cf_handle_ptr = nullptr;
     const rocksdb::Status s =
-        rdb->CreateColumnFamily(opts, cf_name, &cf_handle);
+        rdb->CreateColumnFamily(opts, cf_name, &cf_handle_ptr);
 
     if (s.ok()) {
-      m_cf_name_map[cf_handle->GetName()] = cf_handle;
-      m_cf_id_map[cf_handle->GetID()] = cf_handle;
-    } else {
-      cf_handle = nullptr;
+      DBUG_ASSERT(cf_handle_ptr != nullptr);
+      cf_handle.reset(cf_handle_ptr);
+      m_cf_name_map[cf_handle_ptr->GetName()] = cf_handle;
+      m_cf_id_map[cf_handle_ptr->GetID()] = cf_handle;
     }
   }
 
@@ -133,18 +135,25 @@ rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_or_create_cf(
 /*
   Find column family by its cf_name.
 */
+std::shared_ptr<rocksdb::ColumnFamilyHandle> Rdb_cf_manager::get_cf(
+    const std::string &cf_name) const {
+  return get_cf(cf_name, false /*lock_held_by_caller*/);
+}
 
-rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_cf(
-    const std::string &cf_name_arg, const bool lock_held_by_caller) const {
-  rocksdb::ColumnFamilyHandle *cf_handle;
+std::shared_ptr<rocksdb::ColumnFamilyHandle> Rdb_cf_manager::get_cf(
+    const std::string &cf_name, const bool lock_held_by_caller) const {
+  DBUG_ASSERT(!cf_name.empty());
+  std::shared_ptr<rocksdb::ColumnFamilyHandle> cf_handle;
 
   if (!lock_held_by_caller) {
     RDB_MUTEX_LOCK_CHECK(m_mutex);
   }
-  std::string cf_name = cf_name_arg.empty() ? DEFAULT_CF_NAME : cf_name_arg;
 
   const auto it = m_cf_name_map.find(cf_name);
-  cf_handle = (it != m_cf_name_map.end()) ? it->second : nullptr;
+
+  if (it != m_cf_name_map.end()) {
+    cf_handle = it->second;
+  }
 
   if (!cf_handle) {
     // NO_LINT_DEBUG
@@ -158,8 +167,9 @@ rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_cf(
   return cf_handle;
 }
 
-rocksdb::ColumnFamilyHandle *Rdb_cf_manager::get_cf(const uint32_t id) const {
-  rocksdb::ColumnFamilyHandle *cf_handle = nullptr;
+std::shared_ptr<rocksdb::ColumnFamilyHandle> Rdb_cf_manager::get_cf(
+    const uint32_t id) const {
+  std::shared_ptr<rocksdb::ColumnFamilyHandle> cf_handle;
 
   RDB_MUTEX_LOCK_CHECK(m_mutex);
   const auto it = m_cf_id_map.find(id);
@@ -181,9 +191,9 @@ std::vector<std::string> Rdb_cf_manager::get_cf_names(void) const {
   return names;
 }
 
-std::vector<rocksdb::ColumnFamilyHandle *> Rdb_cf_manager::get_all_cf(
-    void) const {
-  std::vector<rocksdb::ColumnFamilyHandle *> list;
+std::vector<std::shared_ptr<rocksdb::ColumnFamilyHandle>>
+Rdb_cf_manager::get_all_cf(void) const {
+  std::vector<std::shared_ptr<rocksdb::ColumnFamilyHandle>> list;
 
   RDB_MUTEX_LOCK_CHECK(m_mutex);
 
@@ -197,66 +207,56 @@ std::vector<rocksdb::ColumnFamilyHandle *> Rdb_cf_manager::get_all_cf(
   return list;
 }
 
-struct Rdb_cf_scanner : public Rdb_tables_scanner {
-  uint32_t m_cf_id;
-  int m_is_cf_used;
-
-  explicit Rdb_cf_scanner(uint32_t cf_id)
-      : m_cf_id(cf_id), m_is_cf_used(false) {}
-
-  int add_table(Rdb_tbl_def *tdef) override {
-    DBUG_ASSERT(tdef != nullptr);
-
-    for (uint i = 0; i < tdef->m_key_count; i++) {
-      const Rdb_key_def &kd = *tdef->m_key_descr_arr[i];
-
-      if (kd.get_cf()->GetID() == m_cf_id) {
-        m_is_cf_used = true;
-        return HA_EXIT_SUCCESS;
-      }
-    }
-    return HA_EXIT_SUCCESS;
-  }
-};
-
-int Rdb_cf_manager::drop_cf(const std::string &cf_name) {
-  auto ddl_manager = rdb_get_ddl_manager();
-  uint32_t cf_id = 0;
-
-  if (cf_name == DEFAULT_SYSTEM_CF_NAME) {
-    return HA_EXIT_FAILURE;
-  }
-
+int Rdb_cf_manager::remove_dropped_cf(Rdb_dict_manager *const dict_manager,
+                                      rocksdb::TransactionDB *const rdb,
+                                      const uint32 &cf_id) {
+  dict_manager->assert_lock_held();
   RDB_MUTEX_LOCK_CHECK(m_mutex);
-  auto cf_handle = get_cf(cf_name, true /* lock_held_by_caller */);
-  if (cf_handle == nullptr) {
-    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
-    return HA_EXIT_SUCCESS;
-  }
+  const std::unique_ptr<rocksdb::WriteBatch> wb = dict_manager->begin();
+  rocksdb::WriteBatch *const batch = wb.get();
 
-  cf_id = cf_handle->GetID();
-  Rdb_cf_scanner scanner(cf_id);
-
-  auto ret = ddl_manager->scan_for_tables(&scanner);
-  if (ret) {
+  const auto it = m_cf_id_map.find(cf_id);
+  if (it == m_cf_id_map.end()) {
+    dict_manager->delete_dropped_cf_and_flags(batch, cf_id);
+    dict_manager->commit(batch);
     RDB_MUTEX_UNLOCK_CHECK(m_mutex);
-    return ret;
-  }
 
-  if (scanner.m_is_cf_used) {
-    // column family is used by existing key
-    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+    // NO_LINT_DEBUG
+    sql_print_warning(
+        "RocksDB: Column family with id %u is marked as dropped, "
+        "but doesn't exist in cf manager",
+        cf_id);
+
     return HA_EXIT_FAILURE;
   }
 
-  auto rdb = rdb_get_rocksdb_db();
-  auto status = rdb->DropColumnFamily(cf_handle);
-  if (!status.ok()) {
+  auto cf_handle = it->second.get();
+  const std::string cf_name = cf_handle->GetName();
+
+  if (!dict_manager->get_dropped_cf(cf_id)) {
     RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+    // NO_LINT_DEBUG
+    sql_print_warning(
+        "RocksDB: Column family %s with id %u is not in "
+        "the list of cf ids to be dropped",
+        cf_name.c_str(), cf_id);
+    return HA_EXIT_FAILURE;
+  }
+
+  auto status = rdb->DropColumnFamily(cf_handle);
+
+  if (!status.ok()) {
+    dict_manager->delete_dropped_cf(batch, cf_id);
+    dict_manager->commit(batch);
+    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+
+    // NO_LINT_DEBUG
+    sql_print_error(
+        "RocksDB: Dropping column family %s with id %u on RocksDB failed",
+        cf_name.c_str(), cf_id);
+
     return ha_rocksdb::rdb_error_to_mysql(status);
   }
-
-  delete cf_handle;
 
   auto id_iter = m_cf_id_map.find(cf_id);
   DBUG_ASSERT(id_iter != m_cf_id_map.end());
@@ -266,8 +266,141 @@ int Rdb_cf_manager::drop_cf(const std::string &cf_name) {
   DBUG_ASSERT(name_iter != m_cf_name_map.end());
   m_cf_name_map.erase(name_iter);
 
+  dict_manager->delete_dropped_cf_and_flags(batch, cf_id);
+
+  dict_manager->commit(batch);
   RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+
+  // NO_LINT_DEBUG
+  sql_print_information(
+      "RocksDB: Column family %s with id %u has been dropped successfully",
+      cf_name.c_str(), cf_id);
 
   return HA_EXIT_SUCCESS;
 }
+
+struct Rdb_cf_scanner : public Rdb_tables_scanner {
+  uint32_t m_cf_id;
+
+  explicit Rdb_cf_scanner(uint32_t cf_id) : m_cf_id(cf_id) {}
+
+  int add_table(Rdb_tbl_def *tdef) override {
+    DBUG_ASSERT(tdef != nullptr);
+
+    for (uint i = 0; i < tdef->m_key_count; i++) {
+      const Rdb_key_def &kd = *tdef->m_key_descr_arr[i];
+
+      if (kd.get_cf()->GetID() == m_cf_id) {
+        return HA_EXIT_FAILURE;
+      }
+    }
+    return HA_EXIT_SUCCESS;
+  }
+};
+
+int Rdb_cf_manager::drop_cf(Rdb_ddl_manager *const ddl_manager,
+                            Rdb_dict_manager *const dict_manager,
+                            const std::string &cf_name) {
+  DBUG_ASSERT(!cf_name.empty());
+  dict_manager->assert_lock_held();
+  uint32_t cf_id = 0;
+
+  if (cf_name == DEFAULT_SYSTEM_CF_NAME || cf_name == DEFAULT_CF_NAME) {
+    return HA_EXIT_FAILURE;
+  }
+
+  // We should have already acquired dict manager lock. The order
+  // of lock acquisition here is:
+  //  dict_manager -> cf_manager -> ddl_manager
+  RDB_MUTEX_LOCK_CHECK(m_mutex);
+  auto cf_handle = get_cf(cf_name, true /* lock_held_by_caller */).get();
+  if (cf_handle == nullptr) {
+    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+    // NO_LINT_DEBUG
+    sql_print_warning(
+        "RocksDB: Cannot mark Column family %s to be dropped, "
+        "because it doesn't exist in cf manager",
+        cf_name.c_str());
+
+    return HA_EXIT_FAILURE;
+  }
+
+  cf_id = cf_handle->GetID();
+  Rdb_cf_scanner scanner(cf_id);
+
+  auto ret = ddl_manager->scan_for_tables(&scanner);
+  if (ret) {
+    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+    // NO_LINT_DEBUG
+    sql_print_warning(
+        "RocksDB: Cannot mark Column family %s with id %u to be dropped, "
+        "because it is in use",
+        cf_name.c_str(), cf_id);
+    return ret;
+  }
+
+  ret = ddl_manager->find_in_uncommitted_keydef(cf_id);
+  if (ret) {
+    RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+    // NO_LINT_DEBUG
+    sql_print_warning(
+        "RocksDB: Cannot mark Column family %s with id %u to be dropped, "
+        "because it is used by an ongoing add index command",
+        cf_name.c_str(), cf_id);
+    return ret;
+  }
+
+  // We don't call DropColumnFamily on RocksDB here and
+  // we don't delete handle object. Here we mark the column family
+  // as dropped.
+
+  const std::unique_ptr<rocksdb::WriteBatch> wb = dict_manager->begin();
+  rocksdb::WriteBatch *const batch = wb.get();
+
+  dict_manager->add_dropped_cf(batch, cf_id);
+  dict_manager->commit(batch);
+
+  RDB_MUTEX_UNLOCK_CHECK(m_mutex);
+
+  // NO_LINT_DEBUG
+  sql_print_information(
+      "RocksDB: Column family %s with id %u has been marked to be dropped",
+      cf_name.c_str(), cf_id);
+
+  return HA_EXIT_SUCCESS;
+}
+
+int Rdb_cf_manager::create_cf_flags_if_needed(
+    const Rdb_dict_manager *const dict_manager, const uint32 &cf_id,
+    const std::string &cf_name, const bool is_per_partition_cf) {
+  DBUG_ASSERT(!cf_name.empty());
+  uchar flags =
+      (is_cf_name_reverse(cf_name.c_str()) ? Rdb_key_def::REVERSE_CF_FLAG : 0) |
+      (is_per_partition_cf ? Rdb_key_def::PER_PARTITION_CF_FLAG : 0);
+
+  uint existing_cf_flags;
+  if (dict_manager->get_cf_flags(cf_id, &existing_cf_flags)) {
+    // For the purposes of comparison we'll clear the partitioning bit. The
+    // intent here is to make sure that both partitioned and non-partitioned
+    // tables can refer to the same CF.
+    existing_cf_flags &= ~Rdb_key_def::CF_FLAGS_TO_IGNORE;
+    flags &= ~Rdb_key_def::CF_FLAGS_TO_IGNORE;
+
+    if (existing_cf_flags != flags) {
+      my_error(ER_CF_DIFFERENT, MYF(0), cf_name.c_str(), flags,
+               existing_cf_flags);
+
+      return HA_EXIT_FAILURE;
+    }
+  } else {
+    const std::unique_ptr<rocksdb::WriteBatch> wb = dict_manager->begin();
+    rocksdb::WriteBatch *const batch = wb.get();
+
+    dict_manager->add_cf_flags(batch, cf_id, flags);
+    dict_manager->commit(batch);
+  }
+
+  return HA_EXIT_SUCCESS;
+}
+
 }  // namespace myrocks
