@@ -47,7 +47,8 @@ my_core::PSI_mutex_key rdb_psi_open_tbls_mutex_key, rdb_signal_bg_psi_mutex_key,
     rdb_signal_drop_idx_psi_mutex_key, rdb_signal_mc_psi_mutex_key,
     rdb_collation_data_mutex_key, rdb_mem_cmp_space_mutex_key,
     key_mutex_tx_list, rdb_sysvars_psi_mutex_key, rdb_cfm_mutex_key,
-    rdb_sst_commit_key, rdb_block_cache_resize_mutex_key;
+    rdb_sst_commit_key, rdb_block_cache_resize_mutex_key,
+    rdb_bottom_pri_background_compactions_resize_mutex_key;
 
 my_core::PSI_mutex_info all_rocksdb_mutexes[] = {
     {&rdb_psi_open_tbls_mutex_key, "open tables", PSI_FLAG_GLOBAL},
@@ -63,6 +64,8 @@ my_core::PSI_mutex_info all_rocksdb_mutexes[] = {
     {&rdb_sst_commit_key, "sst commit", PSI_FLAG_GLOBAL},
     {&rdb_block_cache_resize_mutex_key, "resizing block cache",
      PSI_FLAG_GLOBAL},
+    {&rdb_bottom_pri_background_compactions_resize_mutex_key,
+     "resizing bottom pri compaction threads", PSI_FLAG_GLOBAL}
 };
 
 my_core::PSI_rwlock_key key_rwlock_collation_exception_list,
